+++ R/MethodsFit.R
@@ -23,21 +23,26 @@ ppc <- function(fit, data, fun = bayesplot::ppc_dens_overlay, ...) {
 #' @description Creates plots where each observation unit has a separate panel.
 #' @inheritParams get_draws
 #' @inheritParams create_plot_df
+#' @inheritParams plot_sim
 #' @param draws see the \code{draws} argument of \code{\link{get_f}}
-#' @param ... keyword arguments to \code{\link{plot_panel}}
 #' @return a \code{ggplot} object
 #' @family model fit visualization functions
-plot_fit <- function(fit, x = "age", group_by = "id", draws = NULL, ...) {
+plot_fit <- function(fit, x = "age", group_by = "id", draws = NULL,
+                     componentwise = FALSE, ...) {
   df_data <- create_plot_df(fit, x, group_by)
   DF <- plot_fit_helper(fit, df_data, draws)
-  h <- plot_panel(
-    df_data = df_data,
-    df_fit =  dollar(DF, "df_fit"),
-    df_ribbon = dollar(DF, "df_ribbon"),
-    fit_alpha = dollar(DF, "fit_alpha"),
-    ...
-  )
-  h <- h + ggplot2::ggtitle("Model fit", subtitle = DF$info)
+  if (componentwise) {
+    stop("not implemented")
+  } else {
+    h <- plot_api_g(
+      df_data = df_data,
+      df_fit =  dollar(DF, "df_fit"),
+      df_fit_err = dollar(DF, "df_ribbon"),
+      fit_alpha = dollar(DF, "fit_alpha"),
+      ...
+    )
+    h <- h + ggplot2::ggtitle("Model fit", subtitle = DF$info)
+  }
   return(h)
 }