+++ R/utils_data.R
@@ -324,3 +324,41 @@ data_types <- function(data) {
   names(types) <- nams
   as.list(types)
 }
+
+#' Add a crossing of two factors to a data frame
+#'
+#' @param df a data frame
+#' @param fac1 name of first factor, must be found in \code{df}
+#' @param fac2 name of second factor, must be found in \code{df}
+#' @param new_name name of the new factor
+#' @return a data frame
+#' @family data utilities
+add_factor_crossing <- function(df, fac1, fac2, new_name) {
+  a <- dollar(df, fac1)
+  b <- dollar(df, fac2)
+  check_not_null(new_name)
+  check_type(a, "factor")
+  check_type(b, "factor")
+  df[[new_name]] <- interaction(a, b, sep = "*")
+  return(df)
+}
+
+#' Data frame and additional information to long format
+#'
+#' @param df a data frame in wide format
+#' @return a data frame where first column is a factor that determines the
+#' column variable name in the original data and second column contains
+#' the actual values
+#' @family data utilities
+to_long_format <- function(df) {
+  nam <- colnames(df)
+  x <- c()
+  J <- ncol(df)
+  N <- nrow(df)
+  for (j in seq_len(J)) {
+    x <- c(x, as.vector(df[, j]))
+  }
+  fac <- rep(nam, each = N)
+  out <- data.frame(fac, x)
+  return(out)
+}